requirements.txt
pandas
misskey
python-dotenv